+++ nekoyume/Assets/Resources/UI/Prefabs/UI_MimisbrunnrPreparation.prefab
@@ -2813,7 +2813,7 @@ GameObject:
   - component: {fileID: 5751051645145681185}
   - component: {fileID: 8826747835552238131}
   - component: {fileID: 6722949989391504744}
-  - component: {fileID: 5032018281988522220}
+  - component: {fileID: 9155559324301978292}
   m_Layer: 5
   m_Name: EquipSlotGlow
   m_TagString: Untagged
@@ -2878,7 +2878,7 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &5032018281988522220
+--- !u!114 &9155559324301978292
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -2887,21 +2887,29 @@ MonoBehaviour:
   m_GameObject: {fileID: 4269160599572352784}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ea01e3caa21d0264cab5f5afa838df19, type: 3}
+  m_Script: {fileID: 11500000, guid: b0e2d7e1f5f0fcf4aa7cc90aaf97dfe6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  StartWithPlay: 1
-  StartDelay: 0
-  Duration: 1
-  TweenType_: 4
-  Ease_: 1
-  CompleteMethod: 
-  ComponentIndex: 0
-  Target: {fileID: 0}
-  CompleteDelay: 0
-  BeginValue: 0.5
-  EndValue: 1
-  ClearOnStop: 0
+  startWithPlay: 1
+  startDelay: 0
+  duration: 1
+  tweenType: 0
+  ease: 1
+  completeMethod: 
+  componentIndex: 0
+  target: {fileID: 0}
+  completeDelay: 0
+  useCustomEaseCurve: 0
+  customEaseCurve:
+    serializedVersion: 2
+    m_Curve: []
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
+  beginValue: 0.5
+  endValue: 1
+  infiniteLoop: 0
+  loopType: 1
 --- !u!1 &5065808764976370009
 GameObject:
   m_ObjectHideFlags: 0
@@ -3489,7 +3497,7 @@ SpriteRenderer:
   m_Size: {x: 0.89, y: 0.39}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
-  m_WasSpriteAssigned: 1
+  m_WasSpriteAssigned: 0
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
 --- !u!1 &6387715629849393821
@@ -3503,7 +3511,7 @@ GameObject:
   - component: {fileID: 7775912007501706911}
   - component: {fileID: 5064747420105962615}
   - component: {fileID: 7171152840029433570}
-  - component: {fileID: 8176425295563055976}
+  - component: {fileID: 2089277455597303416}
   m_Layer: 5
   m_Name: EquipSlotGlow
   m_TagString: Untagged
@@ -3568,7 +3576,7 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &8176425295563055976
+--- !u!114 &2089277455597303416
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -3577,21 +3585,29 @@ MonoBehaviour:
   m_GameObject: {fileID: 6387715629849393821}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ea01e3caa21d0264cab5f5afa838df19, type: 3}
+  m_Script: {fileID: 11500000, guid: b0e2d7e1f5f0fcf4aa7cc90aaf97dfe6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  StartWithPlay: 1
-  StartDelay: 0
-  Duration: 1
-  TweenType_: 4
-  Ease_: 1
-  CompleteMethod: 
-  ComponentIndex: 0
-  Target: {fileID: 0}
-  CompleteDelay: 0
-  BeginValue: 0.5
-  EndValue: 1
-  ClearOnStop: 0
+  startWithPlay: 1
+  startDelay: 0
+  duration: 1
+  tweenType: 0
+  ease: 1
+  completeMethod: 
+  componentIndex: 0
+  target: {fileID: 0}
+  completeDelay: 0
+  useCustomEaseCurve: 0
+  customEaseCurve:
+    serializedVersion: 2
+    m_Curve: []
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
+  beginValue: 0.5
+  endValue: 1
+  infiniteLoop: 0
+  loopType: 1
 --- !u!1 &6539034910410575573
 GameObject:
   m_ObjectHideFlags: 0
