--- nekoyume/Assets/Resources/UI/Prefabs/UI_QuestPreparation 1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 334502c565fe545449a20cd96d0002d2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
