--- nekoyume/Assets/_Scripts/UI/Tween/DOTweenImageAlpha.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea01e3caa21d0264cab5f5afa838df19
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
